Add NativeCapture.ts
