+++ Modules/@babylonjs/react-native/index.ts
@@ -1,2 +1,3 @@
 export * from './EngineView';
 export * from './EngineHook';
+export * from './NativeCapture';
